+ no longer print SKEW line to nfs.fb files + add issquare and ispow to function list + cleanup of siqs tdiv code: now requires SSE2 + added perfect power checks to siqs and nfs + on NFS restart, check any rels found against min_rels instead of always 	proceeding to filtering + support for large nfs jobs on windows (file size limitations removed) + fix for the "skipping blank line" infinite loop sometimes encountered during 	-batchfile runs + fixed estimation when allocating memory in SoE wrapper - now gives better 	estimates and saves memory + fixed bug introduced in 1.28.4 where rho() doesn't detect PRP factors correctly 	(thanks wblipp!) + if gnfs-lasieve binaries are not detected prior to starting an nfs job that 	requires sieving, siqs is started instead of aborting (and a warning 	message is printed to the screen and logfile) + added a check for a valid siqs savefile and restart siqs inside of factor()

+++ trunk/top/eratosthenes/primes.c
@@ -29,7 +29,8 @@ void primes_from_lineflags(thread_soedata_t *t)
 
 	count_line(t);
 	num_alloc = t->linecount;
-	ddata->primes = (uint64 *)malloc(num_alloc * sizeof(uint64));
+	//printf("attempting to allocate room for %" PRIu64 " primes based on linecount\n",num_alloc);
+	ddata->primes = (uint64 *)malloc((size_t)(num_alloc * sizeof(uint64)));
 	if (ddata->primes == NULL)
 	{
 		printf("failed to allocate primes array in primes_from_lineflags\n");
@@ -50,8 +51,13 @@ void primes_from_lineflags(thread_soedata_t *t)
 				//only store the prime if it is within our requested bounds
 				if ((prime >= sdata->orig_llimit) &&  (prime <= sdata->orig_hlimit))
 				{
+					//if (it >= num_alloc)
+					//{
+					//	printf("error: did not allocate enough primes!\n");
+					//	fflush(stdout);
+					//}
 					ddata->primes[it] = prime;
-					it++;
+					it++;					
 				}
 				//else
 					//printf("prime %" PRIu64 " is outside of original limits of %" PRIu64 " and %" PRIu64"\n",
